# [probably don't merge this] Faster ZKEdDSAEventTicketPCD initialization