As part of proving/verifying, ZKEdDSAEventTicketPCD calls `ensureDepsInitialized`. I've timed this as taking about 2 seconds on my MBA M1.

However, most of the work being done here is repetition!

We call `buildBabyJub()`, which calls `getCurveFromName("bn128", true)`, which takes around 300ms. But then we call `buildEddsa`, which _also_ calls `buildBabyJub()`. Even worse, `buildEddsa()` calls other functions that also call `getCurveFromName("bn128", true)`, each time taking 300ms to do something that has already been done.

Unfortunately there's no API that allows for a more sensible approach, so we have to patch `circomlibjs`. This PR takes a simple approach to patching the library and its type definitions, with the effect that `ensureDepsInitialized` ends up taking about 800ms in total, rather than 2000ms. I think this is pretty important as this is required before we can generate ZK QR codes (required for #772) and also before anyone can use `ZKEdDSAEventTicketPCD` for authentication (if that is the approach we go with).